Added config files and refactored project by creating route folder
