+++ config/README.txt
@@ -1,0 +1,2 @@
+In reall projects do not upload this to version control system. 
+All the config files needs to be ignored while updating repository.
