+++ config/custom-environment-variables.json
@@ -1,0 +1,5 @@
+{
+    "mail":{
+        "password":"app_password"
+    }
+}
