+++ config/default.json
@@ -1,0 +1,3 @@
+{
+    "name":"Vidly App"
+}
