+++ config/production.json
@@ -1,0 +1,6 @@
+{
+    "name":"Vidly App - Production",
+    "mail":{
+        "host":"prod-mail-sever"
+    }
+}
