+++ package.json
@@ -10,6 +10,7 @@
   "author": "H.V. Yohan",
   "license": "ISC",
   "dependencies": {
+    "config": "^3.3.6",
     "express": "^4.17.1",
     "helmet": "^4.5.0",
     "joi": "^17.4.0",
